+++ frontend/src/App.tsx
@@ -8,6 +8,7 @@ import { Link, Route, Routes } from 'react-router-dom';
 import Home from './pages/Home';
 import Login from './pages/login/Login';
 import Dashboard from './pages/Dashboard';
+import OAuth from './pages/oauth/Oauth';
 
 export interface PageProps {
   client: AppClient;
@@ -62,10 +63,8 @@ export function App() {
             <Routes>
               <Route path="/" element={<Home />} />
               <Route path="/login" element={<Login clientID={clientId} />} />
-              <Route
-                path="/dashboard"
-                element={<Dashboard client={client} />}
-              />
+              <Route path="/dashboard" element={<Dashboard />} />
+              <Route path="/oauth" element={<OAuth client={client} />} />
             </Routes>
           </>
         ) : (
